Cambiar columna track de INTEGER a FLOAT para soportar rumbos decimales

+++ db.sql
@@ -12,7 +12,7 @@ CREATE TABLE reports
     lon     FLOAT NULL,
     altitude     FLOAT NULL, -- m
     vert_rate     FLOAT NULL, -- m
-    track     INTEGER NULL,
+    track     FLOAT NULL,
     speed     FLOAT NULL, -- mps
     seen_at     TIMESTAMP NULL,
     messages     INTEGER NULL,
